Factored out analyser types and implemented the actual analysis.
