+++ src/analyser/mod.rs
@@ -1,184 +1,20 @@
-use std::iter::FromIterator;
-
+use Model;
 use errors::*;
 use ops::Op;
-use tfpb::types::DataType;
-use Matrix;
+use Plan;
+
+mod types;
+pub use self::types::*;
 
 #[macro_use]
 pub mod macros;
+
 #[macro_use]
 pub mod helpers;
 
 #[cfg(test)]
 mod tests;
 
-/// An abstract tensor.
-///
-/// The task of the analyser is to tag every edge in the graph with information
-/// about the tensors that flow through it - specifically their datatype, their
-/// shape and possibly their value. During the analysis, however, we might only
-/// know some of that information (say, for instance, that an edge only carries
-/// tensors of rank 4, but without knowing their precise dimension).
-///
-/// This is where abstract tensors come in: they hold partial information about
-/// the datatype, shape and value of tensors that might flow through an edge of
-/// the graph. The analyser will first tag each edge of the graph with the most
-/// general abstract tensor possible, and after each iteration of the analysis,
-/// the tensors will become more and more specialized - until reaching a fixed
-/// point that will hopefully contain enough information for us to work with.
-#[derive(Debug, Clone, PartialEq)]
-pub struct ATensor {
-    pub datatype: AType,
-    pub shape: AShape,
-    pub value: AValue,
-}
-
-impl ATensor {
-    /// Constructs a new abstract tensor, which is as general as possible.
-    pub fn new() -> ATensor {
-        ATensor {
-            datatype: AType::Any,
-            shape: AShape::any(),
-            value: AValue::Any,
-        }
-    }
-}
-
-/// An abstract type.
-#[derive(Debug, Clone, PartialEq)]
-pub enum AType {
-    Any,
-    Only(DataType),
-}
-
-/// An abstract shape.
-/// They are used to represent partial information about the shapes of tensors.
-///
-/// A basic example of abstract shape is `ashape![1, 2]` - which corresponds to
-/// the shape `[1, 2]` in Tensorflow. We can use unknown dimensions in abstract
-/// shapes: `ashape![1, 2, _]` corresponds to any shape `[1, 2, k]`, with `k` a
-/// nonnegative integer. We can also use `..` to only describe the beginning of
-/// a shape, so `ashape![1; ..]` matches any shape that starts with `[1]` (e.g.
-/// `[1]`, `[1, k]`, etc.), and `ashape![..]` matches any shape.
-#[derive(Debug, Clone, PartialEq)]
-pub enum AShape {
-    Open(Vec<ADimension>),
-    Closed(Vec<ADimension>),
-}
-
-impl AShape {
-    /// Returns the most general abstract shape possible.
-    pub fn any() -> AShape {
-        AShape::Open(vec![])
-    }
-
-    /// Returns whether the abstract shape is open.
-    pub fn is_open(self: &AShape) -> bool {
-        match self {
-            AShape::Open(_) => true,
-            AShape::Closed(_) => false,
-        }
-    }
-
-    /// Returns the vector of dimensions defining the abstract shape.
-    pub fn inner(self: &AShape) -> &Vec<ADimension> {
-        match self {
-            AShape::Open(v) | AShape::Closed(v) => v,
-        }
-    }
-
-    /// Tries to transform the abstract shape into a Vec<usize>, or returns
-    /// an Err if some of the dimensions are unknown.
-    pub fn concretize(self: &AShape) -> Result<Vec<usize>> {
-        match self {
-            AShape::Open(_) =>
-                bail!("Impossible to concretize an open shape."),
-            AShape::Closed(v) => v
-                .iter()
-                .map(|d| match d {
-                    ADimension::Any =>
-                        bail!("Impossible to concretize a shape with an unknown dimension."),
-                    ADimension::Only(i) =>
-                        Ok(*i)
-                })
-                .collect()
-        }
-    }
-}
-
-impl FromIterator<usize> for AShape {
-    /// Converts an iterator over usize into a closed shape.
-    fn from_iter<I: IntoIterator<Item=usize>>(iter: I) -> AShape {
-        AShape::Closed(iter
-            .into_iter()
-            .map(|d| ADimension::Only(d))
-            .collect())
-    }
-}
-
-impl<'a> FromIterator<&'a usize> for AShape {
-    /// Converts an iterator over &usize into a closed shape.
-    fn from_iter<I: IntoIterator<Item=&'a usize>>(iter: I) -> AShape {
-        AShape::Closed(iter
-            .into_iter()
-            .map(|d| ADimension::Only(*d))
-            .collect())
-    }
-}
-
-impl<'a> From<&'a[usize]> for AShape {
-    /// Converts an usize slice into a closed shape.
-    fn from(slice: &'a[usize]) -> AShape {
-        slice.iter().collect()
-    }
-}
-
-/// An abstract dimension.
-#[derive(Debug, Clone, PartialEq)]
-pub enum ADimension {
-    Any,
-    Only(usize),
-}
-
-impl ADimension {
-    /// Returns whether the dimension is concrete.
-    pub fn is_concrete(&self) -> bool {
-        match self {
-            ADimension::Any => false,
-            ADimension::Only(_) => true
-        }
-    }
-}
-
-/// An abstract value.
-#[derive(Debug, Clone, PartialEq)]
-pub enum AValue {
-    Any,
-    Only(Matrix),
-}
-
-impl AValue {
-    // Tries to transform the abstract value into a Matrix, or returns an Err.
-    pub fn concretize(self: &AValue) -> Result<&Matrix> {
-        match self {
-            AValue::Any =>
-                bail!("Impossible to concretize an Any value."),
-            AValue::Only(m) =>
-                Ok(m)
-        }
-    }
-
-    // Applies fn to a defined value, and leaves an unknown value untouched.
-    // Returns an Err if something went wrong during the transformation.
-    pub fn map_err<F>(self: &AValue, f: F) -> Result<AValue>
-    where F: Fn(&Matrix) -> Result<Matrix> {
-        match self {
-            AValue::Any => Ok(AValue::Any),
-            AValue::Only(m) => Ok(AValue::Only(f(m)?))
-        }
-    }
-}
 
 /// Attempts to unify two abstract tensors into a more specialized one.
 pub fn unify(x: &ATensor, y: &ATensor) -> Result<ATensor> {
@@ -252,4 +88,86 @@ fn unify_value(x: &AValue, y: &AValue) -> Result<AValue> {
     };
 
     Ok(value)
+}
+
+#[derive(Debug)]
+pub struct Edge {
+    id: usize,
+    from_node: usize,
+    from_out: usize,
+    to_node: usize,
+    tensor: ATensor
+}
+
+/// Runs the analyser on the given graph.
+///
+/// The output argument is used to infer an execution plan for the graph.
+/// Changing it won't alter the correctness of the analysis, but it might
+/// take much longer to complete.
+pub fn analyse<'a>(model: &'a Model, output: usize) -> Result<Vec<Edge>> {
+    // We first give an identity to each edge of the graph.
+    let mut edges = vec![];
+    let mut prev_edges = vec![vec![]; model.nodes().len()];
+    let mut next_edges = vec![vec![]; model.nodes().len()];
+
+    for node in model.nodes() {
+        for input in &node.inputs {
+            let id = edges.len();
+
+            edges.push(Edge {
+                id,
+                from_node: input.0,
+                from_out: input.1.unwrap_or(0),
+                to_node: node.id,
+                tensor: ATensor::new()
+            });
+
+            prev_edges[node.id].push(id);
+            next_edges[input.0].push(id);
+        }
+    }
+
+    // Compute and run an execution plan for the graph.
+    let plan = Plan::for_node(model, output)?;
+    let mut changed;
+
+    macro_rules! one_pass {
+        ($source:ident, $target:ident, $fn:ident) => ({
+            for &n in &plan.order {
+                let inferred = {
+                    let sources: Vec<_> = $source[n].iter().map(|&i| &edges[i].tensor).collect();
+
+                    let inferred = model.get_node_by_id(n)?.op.$fn(sources);
+
+                    if inferred.is_err() {
+                        println!("Error while inferring for {}: {}", n, inferred.unwrap_err());
+                        continue;
+                    }
+
+                    inferred.unwrap()
+                };
+
+                for &j in &$target[n] {
+                    let unified = unify(&inferred[j], &edges[j].tensor)?;
+                    if unified != edges[j].tensor {
+                        edges[j].tensor = unified;
+                        changed = true;
+                    }
+                }
+            }
+        })
+    };
+
+    loop {
+        changed = false;
+
+        one_pass!(prev_edges, next_edges, infer_forward);
+        one_pass!(next_edges, prev_edges, infer_backward);
+
+        if !changed {
+            break;
+        }
+    }
+
+    Ok(edges)
 }
